Move resolver folder into /graphql
