+++ packages/graphql_server/src/graphql/resolver/projects.ts
@@ -1,4 +1,4 @@
-import supabase from '../supabase'
+import supabase from '../../supabase'
 
 export const exposeProjectsData = async () => {
   const { data, error: projectsRetrievalError } = await supabase.from('project').select('*')
