+++ packages/graphql_server/src/graphql/resolvers.ts
@@ -1,4 +1,4 @@
-import { exposeProjectsData } from '../resolver/projects'
+import { exposeProjectsData } from './resolver/projects'
 import { createProject } from '../dbUpdater'
 import { parseGitHubUrl } from '../utils'
